requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-google-genai
python-dotenv
gunicorn
google-cloud-bigquery